requirements.txt
fastapi==0.115.2
python_jose==3.3.0
starlette==0.41.0
